+++ CHANGELOG.md
@@ -4,7 +4,7 @@
   existentialized, to be chosen by the caller.
 
 * BREAKING CHANGE: We don't use `DefaultSignatures` for `flay = gflay` anymore.
-  This is very sud, but unfortunately type inferrence for the `c` type variable
+  This is very sad, but unfortunately type inferrence for the `c` type variable
   in new `Flayable` instances doesn't work.
 
 * Added `Flayable1`, `trivial1`, `collect1`, `Terminal`, `GTerminal`, `zip1`,